chore: created Folders & comp for How It Works page
