+++ scrybe_frontend/src/Pages/HowItWorks/HowItWorks.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function HowItWorks() {
+  return (
+    <div>
+      <h1>how it works</h1>
+    </div>
+  )
+}
+
+export default HowItWorks
