+++ scrybe_frontend/src/Pages/indsutry/industry.jsx
@@ -1,7 +1,8 @@
 import React from "react";
 
 import styles from "../../styles/Industry/industry.module.css";
-import Image from "../../assets/Industry/industryImage.png";
+import IndustryCard from '../../component/Industry/industryCard';
+
 
 function Industry() {
   return (
@@ -16,26 +17,7 @@ function Industry() {
           and products with state-of-the-art AI models.
         </p>
       </div>
-
-      <div className={styles.card}>
-        <div className={styles.titleCard}>
-          <img src={Image} alt="" />
-        </div>
-        <div className={styles.postDetails}>
-          <h2>Industry</h2>
-          <h4>
-            <span>Munaike</span>Wednesday 18th Nov. 2022{" "}
-          </h4>
-          <h3>Why you should embrace Emotional AI - Analytics insights</h3>
-          <p>
-            We often feel that emotions have no place in business but it is
-            thecore of what drives it. Humans are emotional and humans build
-            andmakeup businesses. They&apos;re part of who we are, and it&apos;s
-            important to be able to harness their potential. Emotions have a
-            significant impact on business and your company&apos;s success
-          </p>
-        </div>
-      </div>
+        <IndustryCard/>
     </div>
   );
 }
