--- scrybe_frontend/src/component/industryCard.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function industryCard() {
-  return <div />;
-}
-
-export default industryCard;
